# ai-service/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "江苏 13 市 AI 稽核助手",
    "description": "江苏 13 市 AI 稽核助手 REST API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/audit-tasks": {
      "post": {
        "tags": [
          "audit-tasks"
        ],
        "summary": "Create Audit Task",
        "operationId": "create_audit_task_api_v1_audit_tasks_post",
        "parameters": [
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditTaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditTaskView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "audit-tasks"
        ],
        "summary": "List Audit Tasks",
        "operationId": "list_audit_tasks_api_v1_audit_tasks_get",
        "parameters": [
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditTaskList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit-tasks/{task_id}": {
      "get": {
        "tags": [
          "audit-tasks"
        ],
        "summary": "Get Audit Task",
        "operationId": "get_audit_task_api_v1_audit_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditTaskView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit-cases": {
      "post": {
        "tags": [
          "audit-cases"
        ],
        "summary": "Create Audit Case",
        "description": "同步结构化一份历史报告。\n\n当前材料规模较小，因此接口同步返回201；重复调用会更新同一案例并递增版本。\n将来改为后台批处理时会新增异步入口，不改变现有接口的状态码语义。",
        "operationId": "create_audit_case_api_v1_audit_cases_post",
        "parameters": [
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditCaseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditCaseView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "audit-cases"
        ],
        "summary": "List Audit Cases",
        "description": "查询当前城市可审计的历史案例；X-City-Code由公共依赖统一校验。",
        "operationId": "list_audit_cases_api_v1_audit_cases_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditCaseList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit-tasks/{task_id}/analysis-runs": {
      "post": {
        "tags": [
          "analysis-runs"
        ],
        "summary": "Create Analysis Run",
        "operationId": "create_analysis_run_api_v1_audit_tasks__task_id__analysis_runs_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRunCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis-runs": {
      "get": {
        "tags": [
          "analysis-runs"
        ],
        "summary": "List Analysis Runs",
        "operationId": "list_analysis_runs_api_v1_analysis_runs_get",
        "parameters": [
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis-runs/{run_id}": {
      "get": {
        "tags": [
          "analysis-runs"
        ],
        "summary": "Get Analysis Run",
        "operationId": "get_analysis_run_api_v1_analysis_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis-runs/{run_id}/resume": {
      "post": {
        "tags": [
          "analysis-runs"
        ],
        "summary": "Resume Analysis Run",
        "description": "确认系统是否超标，并从同一PostgreSQL检查点继续原运行。",
        "operationId": "resume_analysis_run_api_v1_analysis_runs__run_id__resume_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "X-City-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-City-Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRunResume"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisEvent": {
        "properties": {
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "node": {
            "type": "string",
            "title": "Node"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "sequence",
          "node",
          "message",
          "created_at"
        ],
        "title": "AnalysisEvent"
      },
      "AnalysisRunCreate": {
        "properties": {
          "material_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Material Refs"
          }
        },
        "type": "object",
        "title": "AnalysisRunCreate",
        "description": "一次分析显式选择的本次任务截图ID。"
      },
      "AnalysisRunList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnalysisRunView"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AnalysisRunList"
      },
      "AnalysisRunResume": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "confirm_over_limit",
              "confirm_not_over_limit"
            ],
            "title": "Decision"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "AnalysisRunResume",
        "description": "恢复“系统是否超标”人工中断时的REST输入。"
      },
      "AnalysisRunView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "city_code": {
            "type": "string",
            "title": "City Code"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "current_node": {
            "type": "string",
            "title": "Current Node"
          },
          "material_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Material Refs"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/AnalysisEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "pending_interrupt": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Interrupt"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "city_code",
          "status",
          "progress",
          "current_node",
          "material_refs",
          "events",
          "pending_interrupt",
          "result",
          "error",
          "created_by",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "AnalysisRunView"
      },
      "AuditCaseCreate": {
        "properties": {
          "source_document_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 36,
            "title": "Source Document Id"
          }
        },
        "type": "object",
        "required": [
          "source_document_id"
        ],
        "title": "AuditCaseCreate",
        "description": "指定要结构化的历史报告。"
      },
      "AuditCaseList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditCaseView"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AuditCaseList"
      },
      "AuditCaseView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "city_code": {
            "type": "string",
            "title": "City Code"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "source_document_id": {
            "type": "string",
            "title": "Source Document Id"
          },
          "source_title": {
            "type": "string",
            "title": "Source Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "memory_status": {
            "type": "string",
            "title": "Memory Status"
          },
          "billing_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Period"
          },
          "over_limit_items": {
            "items": {},
            "type": "array",
            "title": "Over Limit Items"
          },
          "primary_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Reason"
          },
          "reason_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Category"
          },
          "key_facts": {
            "items": {},
            "type": "array",
            "title": "Key Facts"
          },
          "evidence_element_ids": {
            "items": {},
            "type": "array",
            "title": "Evidence Element Ids"
          },
          "uncertain_items": {
            "items": {},
            "type": "array",
            "title": "Uncertain Items"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "analyzed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyzed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "city_code",
          "site_id",
          "source_document_id",
          "source_title",
          "status",
          "memory_status",
          "billing_period",
          "over_limit_items",
          "primary_reason",
          "reason_category",
          "key_facts",
          "evidence_element_ids",
          "uncertain_items",
          "confidence",
          "model_name",
          "prompt_version",
          "error_code",
          "error_message",
          "created_at",
          "analyzed_at"
        ],
        "title": "AuditCaseView",
        "description": "供检索审计和前端查看的案例资源。\n\n响应包含原报告ID与证据元素ID，但不暴露storage_key和服务器绝对路径。"
      },
      "AuditTaskCreate": {
        "properties": {
          "site_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 36,
            "title": "Site Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Title"
          },
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 2,
            "title": "Question",
            "default": "请分析本次电费超标情况，并与本市历史原因进行比较。"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "title"
        ],
        "title": "AuditTaskCreate",
        "description": "创建稽核任务时由前端提交的稳定业务字段。"
      },
      "AuditTaskList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditTaskView"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AuditTaskList"
      },
      "AuditTaskView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "city_code": {
            "type": "string",
            "title": "City Code"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "site_code": {
            "type": "string",
            "title": "Site Code"
          },
          "site_name": {
            "type": "string",
            "title": "Site Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "city_code",
          "site_id",
          "site_code",
          "site_name",
          "title",
          "question",
          "status",
          "created_by",
          "created_at"
        ],
        "title": "AuditTaskView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}